Refactor DatasetsContent to utilize useLayoutActions for improved flow management
